[api-home-and-trails] isolate loadmore, input, search logic
